Changed package name following change of project name
