+++ .gitattributes
@@ -1,1 +1,1 @@
-/src/test/resources/uk/gov/ons/sbr/scat/data/csv/*.csv text eol=crlf
+/src/test/resources/uk/gov/ons/sbr/sit/data/csv/*.csv text eol=crlf
