--- src/main/scala/uk/gov/ons/sbr/scat/url/BaseUrl.scala
@@ -1,8 +1,0 @@
-package uk.gov.ons.sbr.scat.url
-
-case class BaseUrl(protocol: String, host: String, port: Int)
-
-object BaseUrl {
-  def asUrlString(baseUrl: BaseUrl): String =
-    s"${baseUrl.protocol}://${baseUrl.host}:${baseUrl.port}"
-}
