--- src/main/scala/uk/gov/ons/sbr/scat/url/Url.scala
@@ -1,6 +1,0 @@
-package uk.gov.ons.sbr.scat.url
-
-object Url {
-  def apply(withBase: BaseUrl, withPath: String): String =
-    s"${BaseUrl.asUrlString(withBase)}/$withPath"
-}
