--- src/main/scala/uk/gov/ons/sbr/scat/ws/WithActorSystem.scala
@@ -1,11 +1,0 @@
-package uk.gov.ons.sbr.scat.ws
-
-import akka.actor.ActorSystem
-
-object WithActorSystem {
-  def apply[A](f: ActorSystem => A): A = {
-    val actorSystem = ActorSystem()
-    try f(actorSystem)
-    finally actorSystem.terminate()
-  }
-}
