+++ src/main/scala/uk/gov/ons/sbr/sit/data/package.scala
@@ -1,0 +1,9 @@
+package uk.gov.ons.sbr.sit
+
+package object data {
+  type ColumnName = String
+  type Fields = Map[ColumnName, String]
+  type Field = (ColumnName, String)
+  type Row = Fields
+  type ErrorMessage = String
+}
