+++ src/main/scala/uk/gov/ons/sbr/sit/url/Url.scala
@@ -1,0 +1,6 @@
+package uk.gov.ons.sbr.sit.url
+
+object Url {
+  def apply(withBase: BaseUrl, withPath: String): String =
+    s"${BaseUrl.asUrlString(withBase)}/$withPath"
+}
