--- src/test/scala/uk/gov/ons/sbr/scat/it/fixture/IntegrationSpec.scala
@@ -1,6 +1,0 @@
-package uk.gov.ons.sbr.scat.it.fixture
-
-import org.scalatest.concurrent.ScalaFutures
-import org.scalatest.{GivenWhenThen, Matchers, fixture}
-
-trait IntegrationSpec extends fixture.FeatureSpec with GivenWhenThen with Matchers with ScalaFutures
