--- src/test/scala/uk/gov/ons/sbr/scat/it/fixture/SbrControlIntegrationSpec.scala
@@ -1,8 +1,0 @@
-package uk.gov.ons.sbr.scat.it.fixture
-
-import org.scalatest.time.{Millis, Seconds, Span}
-
-trait SbrControlIntegrationSpec extends IntegrationSpec with WSRequestFixture {
-  implicit val defaultPatience: PatienceConfig =
-    PatienceConfig(timeout =  Span(5, Seconds), interval = Span(10, Millis))
-}
